Update Example_Fig_2_calculate.m
I added the code to remove extra files with .html file extensions into the Example_Fig_2_calculate.m

This is needed as when multiple file versions are available during download a download failure occurs and adds an .html file with the data's name. Before calculations happen these lines remove the html files and allows the code to run uninterrupted.